rename modules for consistency
